+++ R/reduce.R
@@ -19,6 +19,10 @@
 #'   function will be passed the accumulated value as the first
 #'   argument, the next value of `.x` as the second argument, and the
 #'   next value of `.y` as the third argument.
+#'
+#'   The reduction terminates early if `.f` returns a value wrapped in
+#'   a [done()].
+#'
 #' @param .init If supplied, will be used as the first value to start
 #'   the accumulation, rather than using `x[[1]]`. This is useful if
 #'   you want to ensure that `reduce` returns a correct value when `.x`
@@ -93,6 +97,29 @@
 #' x <- list(c(0, 1), c(2, 3), c(4, 5))
 #' y <- list(c(6, 7), c(8, 9))
 #' reduce2(x, y, paste)
+#'
+#'
+#' # You can shortcircuit a reduction and terminate it early by
+#' # returning a value wrapped in a done(). In the following example
+#' # we return early if the result-so-far, which is passed on the LHS,
+#' # meets a condition:
+#' paste3 <- function(out, input, sep = ".") {
+#'   if (nchar(out) > 4) {
+#'     return(done(out))
+#'   }
+#'   paste(out, input, sep = sep)
+#' }
+#' letters %>% reduce(paste3)
+#'
+#' # Here the early return branch checks the incoming inputs passed on
+#' # the RHS:
+#' paste4 <- function(out, input, sep = ".") {
+#'   if (input == "j") {
+#'     return(done(out))
+#'   }
+#'   paste(out, input, sep = sep)
+#' }
+#' letters %>% reduce(paste4)
 #' @export
 reduce <- function(.x, .f, ..., .init, .dir = c("forward", "backward")) {
   reduce_impl(.x, .f, ..., .init = .init, .dir = .dir)
@@ -103,11 +130,16 @@ reduce2 <- function(.x, .y, .f, ..., .init) {
   reduce2_impl(.x, .y, .f, ..., .init = .init, .left = TRUE)
 }
 
-reduce_impl <- function(.x, .f, ..., .init, .dir = "forward") {
+reduce_impl <- function(.x, .f, ..., .init, .dir, .acc = FALSE) {
   left <- arg_match(.dir, c("forward", "backward")) == "forward"
 
   out <- reduce_init(.x, .init, left = left)
   idx <- reduce_index(.x, .init, left = left)
+
+  if (.acc) {
+    acc_out <- accum_init(out, idx, left = left)
+    acc_idx <- accum_index(acc_out, left = left)
+  }
 
   .f <- as_mapper(.f, ...)
 
@@ -120,11 +152,48 @@ reduce_impl <- function(.x, .f, ..., .init, .dir = "forward") {
     fn <- function(x, y, ...) .f(y, x, ...)
   }
 
-  for (i in idx) {
-    out <- fn(out, .x[[i]], ...)
+  for (i in seq_along(idx)) {
+    prev <- out
+    elt <- .x[[idx[[i]]]]
+    out <- fn(out, elt, ...)
+
+    if (is_done_box(out)) {
+      return(reduce_early(out, prev, .acc, acc_out, acc_idx[[i]], left))
+    }
+
+    if (.acc) {
+      acc_out[[acc_idx[[i]]]] <- out
+    }
   }
 
-  out
+  if (.acc) {
+    acc_out
+  } else {
+    out
+  }
+}
+
+reduce_early <- function(out, prev, acc, acc_out, acc_idx, left = TRUE) {
+  if (is_done_box(out, empty = TRUE)) {
+    out <- prev
+    offset <- if (left) -1L else 1L
+  } else {
+    out <- unbox(out)
+    offset <- 0L
+  }
+
+  if (!acc) {
+    return(out)
+  }
+
+  acc_idx <- acc_idx + offset
+  acc_out[[acc_idx]] <- out
+
+  if (left) {
+    acc_out[seq_len(acc_idx)]
+  } else {
+    acc_out[seq(acc_idx, length(acc_out))]
+  }
 }
 
 reduce_init <- function(x, init, left = TRUE) {
@@ -158,11 +227,29 @@ reduce_index <- function(x, init, left = TRUE) {
   }
 }
 
-reduce2_impl <- function(.x, .y, .f,
-                         ...,
-                         .init,
-                         .left = TRUE,
-                         .acc = FALSE) {
+accum_init <- function(first, idx, left) {
+  len <- length(idx) + 1L
+  out <- new_list(len)
+
+  if (left) {
+    out[[1]] <- first
+  } else {
+    out[[len]] <- first
+  }
+
+  out
+}
+accum_index <- function(out, left) {
+  n <- length(out)
+
+  if (left) {
+    seq_len2(2, n)
+  } else {
+    rev(seq_len(n - 1L))
+  }
+}
+
+reduce2_impl <- function(.x, .y, .f, ..., .init, .left = TRUE, .acc = FALSE) {
   out <- reduce_init(.x, .init, left = .left)
   x_idx <- reduce_index(.x, .init, left = .left)
   y_idx <- reduce_index(.y, NULL, left = .left)
@@ -174,25 +261,29 @@ reduce2_impl <- function(.x, .y, .f,
   .f <- as_mapper(.f, ...)
 
   if (.acc) {
-    offset <- !missing(.init)
-    res <- new_list(length(x_idx) + offset)
-    res[[1]] <- out
+    acc_out <- accum_init(out, x_idx, left = .left)
+    acc_idx <- accum_index(acc_out, left = .left)
   }
 
   for (i in seq_along(x_idx)) {
+    prev <- out
+
     x_i <- x_idx[[i]]
     y_i <- y_idx[[i]]
 
     out <- .f(out, .x[[x_i]], .y[[y_i]], ...)
 
+    if (is_done_box(out)) {
+      return(reduce_early(out, prev, .acc, acc_out, acc_idx[[i]]))
+    }
+
     if (.acc) {
-      res[[x_i + offset]] <- out
+      acc_out[[acc_idx[[i]]]] <- out
     }
   }
 
   if (.acc) {
-    names(res) <- accumulate_names(names(.x), .init, left = .left)
-    res
+    acc_out
   } else {
     out
   }
@@ -228,6 +319,12 @@ seq_len2 <- function(start, end) {
 #'   and the last element is the final reduced value. In case of a
 #'   right accumulation, this order is reversed.
 #'
+#'   The accumulation terminates early if `.f` returns a value wrapped
+#'   in a [done()]. If the done box is empty, the last value is
+#'   used instead and the result is one element shorter (but always
+#'   includes the initial value, even when terminating at the first
+#'   iteration).
+#'
 #' @section Life cycle:
 #'
 #' `accumulate_right()` is soft-deprecated in favour of the `.dir`
@@ -260,6 +357,42 @@ seq_len2 <- function(start, end) {
 #' letters[1:4] %>% accumulate(paste2)
 #' letters[1:4] %>% accumulate2(c("-", ".", "-"), paste2)
 #'
+#'
+#' # You can shortcircuit an accumulation and terminate it early by
+#' # returning a value wrapped in a done(). In the following example
+#' # we return early if the result-so-far, which is passed on the LHS,
+#' # meets a condition:
+#' paste3 <- function(out, input, sep = ".") {
+#'   if (nchar(out) > 4) {
+#'     return(done(out))
+#'   }
+#'   paste(out, input, sep = sep)
+#' }
+#' letters %>% accumulate(paste3)
+#'
+#' # Note how we get twice the same value in the accumulation. That's
+#' # because we have returned it twice. To prevent this, return an empty
+#' # done box to signal to accumulate() that it should terminate with the
+#' # value of the last iteration:
+#' paste3 <- function(out, input, sep = ".") {
+#'   if (nchar(out) > 4) {
+#'     return(done())
+#'   }
+#'   paste(out, input, sep = sep)
+#' }
+#' letters %>% accumulate(paste3)
+#'
+#' # Here the early return branch checks the incoming inputs passed on
+#' # the RHS:
+#' paste4 <- function(out, input, sep = ".") {
+#'   if (input == "f") {
+#'     return(done())
+#'   }
+#'   paste(out, input, sep = sep)
+#' }
+#' letters %>% accumulate(paste4)
+#'
+#'
 #' # Simulating stochastic processes with drift
 #' \dontrun{
 #' library(dplyr)
@@ -275,15 +408,17 @@ seq_len2 <- function(start, end) {
 #' }
 #' @export
 accumulate <- function(.x, .f, ..., .init, .dir = c("forward", "backward")) {
-  left <- arg_match(.dir, c("forward", "backward")) == "forward"
-
+  .dir <- arg_match(.dir, c("forward", "backward"))
   .f <- as_mapper(.f, ...)
-  f <- function(x, y) {
-    .f(x, y, ...)
-  }
 
-  res <- Reduce(f, .x, init = .init, accumulate = TRUE, right = !left)
-  names(res) <- accumulate_names(names(.x), .init, left = left)
+  res <- reduce_impl(.x, .f, ..., .init = .init, .dir = .dir, .acc = TRUE)
+  names(res) <- accumulate_names(names(.x), .init, .dir)
+
+  # FIXME vctrs: This simplification step is for compatibility with
+  # the `base::Reduce()` implementation in earlier purrr versions
+  if (all(map_int(res, length) == 1L)) {
+    res <- unlist(res, recursive = FALSE)
+  }
 
   res
 }
@@ -293,7 +428,7 @@ accumulate2 <- function(.x, .y, .f, ..., .init) {
   reduce2_impl(.x, .y, .f, ..., .init = .init, .acc = TRUE)
 }
 
-accumulate_names <- function(nms, init, left = left) {
+accumulate_names <- function(nms, init, dir) {
   if (is_null(nms)) {
     return(NULL)
   }
@@ -301,7 +436,7 @@ accumulate_names <- function(nms, init, left = left) {
   if (!missing(init)) {
     nms <- c(".init", nms)
   }
-  if (!left) {
+  if (dir == "backward") {
     nms <- rev(nms)
   }
 
@@ -336,7 +471,7 @@ reduce_right <- function(.x, .f, ..., .init) {
     ""
   ))
   .x <- rev(.x) # Compatibility
-  reduce_impl(.x, .f, ..., .init = .init)
+  reduce_impl(.x, .f, ..., .dir = "forward", .init = .init)
 }
 #' @rdname reduce_right
 #' @export