+++ NEWS.md
@@ -141,6 +141,14 @@ reduce2_right(.x = letters[1:4], .y = paste2, .f = c("-", ".", "-")) # working
 
 ## Minor improvements and fixes
 
+* `reduce()`, `reduce2()`, `accumulate()`, and `accumulate2()` now
+  terminate early when the function returns a value wrapped in a
+  `done()` (#253). When an empty `done()` is returned, the
+  value at the last iteration is returned instead.
+
+* New `done()` function. Use it to wrap a value and signal early
+  termination to a caller.
+
 * New `accumulate2()` function which is to `accumulate()` what
   `reduce2()` is to `reduce()`.